+++ release.py
@@ -1,5 +1,0 @@
-# path.py uses jaraco.packaging.release for releases
-files_with_versions = 'path.py',
-
-# build an sdist and wheel
-dist_commands = 'sdist', 'bdist_wheel'